requirements.txt
requests==2.32.5
Pillow==11.3.0